requirements.txt
langchain==0.0.101
gradio==3.20.1
accelerate
addict
albumentations
basicsr
controlnet-aux
diffusers
einops
imageio
imageio-ffmpeg
invisible-watermark
kornia
numpy
omegaconf
open_clip_torch
openai
opencv-python
prettytable
safetensors
streamlit
test-tube
timm
torchmetrics
# transformers
webdataset
yapf

scikit-learn
peft
icetk
cpm_kernels
git+https://github.com/huggingface/transformers.git
